Added the code to pre and post process the output of the GNSS
preprocessing_binary_parser.py; this code repairs directly the raw binary output of the GNSS. As the output (as of now) the lines are not all the same length and sometimes a new line is initiated when that is not necessary. This code fixes that and makes sure that all lines are the same length and correct.

nmea_post_processing.py; this code removes all duplicate entries and makes sure every line is complete in the .txt file and makes the code exactly as the team needs it to be able to see the error.

NMEA_processing.py; this code takes the NMEA data and transforms it into a .txt file with 4 columns: time, latitude, longitude and altitude. It makes sure the signs are correct (West and South negative). Also, It ensures the starting time is correct.

Binary_processing.py; This code is similar to the NMEA processing. However, it now makes a .txt file with 7 columns as requested by the team; time, x, y, z, vx, vy and vz.  

Binary_updated_processing.py; The Binary output was changed with a extra information to see if a line of data is complete or not. This was incorporated into the this code. Also, the time is processed differently, per request of the team.